requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0

# Deep Learning
torch>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.0.0

# API Server
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0

# Data Collection
requests>=2.31.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
joblib>=1.3.0
